backend/requirements.txt
flask
flask-restful 
flask-sqlalchemy 
flask-cors
mysql-connector-python